Results from first ISP survey
These configurations were created by taking the domains of the most common
ISPs, looking up the mail server configuration on their website, by manually
searching and going to search provider support page and capturing it,
putting it in a standardized form.
Then, writing it as Autoconfig XML files, then validating it with a survey
using a script that connects to these servers using the IMAP, POP3 and SMTP
protocols, and checks whether the server is actually there, and whether its
TLS configuration and TLS certificate is proper.

This work was done by Jeremy Chang, alone, under contract from Ben Bucksch.
This is a "clean room" re-creation of the ISPDB, from scratch, from original
sources, without the original ISPDB created originally by Ben Bucksch in 2008.

It was created based on the following instructions:

1. Go through the list of largest domains
2. For each domain, look up the DNS MX record. Look at the MX hostname. Find who is the hoster. Also find out which clusters they have. E.g. Yahoo and Outlook.com have US and European clusters, and we need to differentiate them. But we want to merge live.de, live.fr, live.it, outlook.de, hotmail.de etc. into the same config, if possible. What matters is the mail *storage* cluster, but you can typically see that from the MX hostnames.
3. Find the correct IMAP, POP3 and SMTP configuration, using the customer help or support documents that they publish. You can typically search for "<providername> thunderbird IMAP hostname" or a similar search. But *important* make sure that your are the official website of the provider. Or that the config hostnames match the provider's main domain (e.g. imap.sekan.net for foo@sekan.net).
4. We need: hostname, port number, TLS or STARTTLS (which one of the two) - each for IMAP, POP3 and SMTP. Also username form ("foo@example.com" (preferred) or "foo" or something else?), and whether an app-specific password is required.
5. Write these into an Excel or LibreOffice Calc sheet or a text file.
6. Make a super simple hackish little script that reads your notes and writes the autoconfig XML file format, as defined by https://benbucksch.github.io/autoconfig-spec/draft-ietf-mailmaint-autoconfig.html
7. Check the configs using a node.js script. Use the IMAP server survey script that you created and used a few months ago, to check the configs. Let it read the autoconfig XML file, using the Mustang code, or using [direct TCP connection](https://www.bucksch.org/1/projects/autoconfiguration/how-to-create-your-own-config-file.html#How_to_probe_mail_servers). Obviously, login will not work, but we should at least be able to connect and capture the `CAPABILITIES` before login. Check that the TLS certificate does not error. Check each server config, including all IMAP, POP3 and SMTP configs listed.
8. Correct the configs based on the survey results.
9. Send me the autoconfig XML files.

But *IMPORTANT*: You are *NOT ALLOWED* to look at [any other] ISP DB.
The entire purpose of the exercise is to re-create the information
from scratch, from original sources.
This is important for legal reasons.
This is a "clean room" re-creation of the DB.